requirements.txt
aws-cdk-lib
cdk-nag
constructs
jwcrypto
requests
